Make LikeBatch concrete (#649)
